Fix most of the clippy lints.

+++ kernel/src/boot_aps.rs
@@ -83,7 +83,7 @@ pub unsafe fn boot_aps(madt: &Madt) {
         bspdone = (end) as *mut u32;
         aprunning = &mut *((end + 4) as *mut AtomicU32);
         *((end + 8) as *mut u32) = cr3_addr;
-        *((end + 16) as *mut u64) = ap_startup_f as u64;
+        *((end + 16) as *mut usize) = ap_startup_f as usize;
         core_local_storage = (end + 24) as *mut u64;
     }
